+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/trigger/GroovyTriggerSupporter.java
@@ -1,7 +1,17 @@
 package com.dwarfeng.fdr.impl.handler.trigger;
 
+import com.dwarfeng.dutil.basic.io.IOUtil;
+import com.dwarfeng.dutil.basic.io.StringOutputStream;
 import com.dwarfeng.fdr.impl.handler.TriggerSupporter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
+import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Component;
+
+import java.io.InputStream;
+import java.nio.charset.StandardCharsets;
 
 /**
  * 使用Groovy脚本的触发器支持器。
@@ -13,6 +23,10 @@ import org.springframework.stereotype.Component;
 public class GroovyTriggerSupporter implements TriggerSupporter {
 
     public static final String SUPPORT_TYPE = "groovy_trigger";
+    private static final Logger LOGGER = LoggerFactory.getLogger(GroovyTriggerSupporter.class);
+
+    @Autowired
+    private ApplicationContext applicationContext;
 
     @Override
     public String provideType() {
@@ -21,7 +35,7 @@ public class GroovyTriggerSupporter implements TriggerSupporter {
 
     @Override
     public String provideLabel() {
-        return "Groovy触发器";
+        return "Groovy过滤器";
     }
 
     @Override
@@ -31,6 +45,19 @@ public class GroovyTriggerSupporter implements TriggerSupporter {
 
     @Override
     public String provideExampleContent() {
-        return "";
+        try {
+            Resource resource = applicationContext.getResource("classpath:groovy/ExampleTriggerProcessor.groovy");
+            String example;
+            try (InputStream sin = resource.getInputStream();
+                 StringOutputStream sout = new StringOutputStream(StandardCharsets.UTF_8, true)) {
+                IOUtil.trans(sin, sout, 4096);
+                sout.flush();
+                example = sout.toString();
+            }
+            return example;
+        } catch (Exception e) {
+            LOGGER.warn("读取文件 classpath:groovy/ExampleTriggerProcessor.groovy 时出现异常", e);
+            return "";
+        }
     }
 }